[클라][서버] 클라, 로그인 서버 패킷 송수신 구현

+++ Game/Client/Renderer2D.cpp
@@ -103,7 +103,7 @@ namespace Renderer2D
 		Renderer::ctx->CreateSolidColorBrush(color, &brush);
 
 		auto size{ Renderer::ctx->GetSize() };
-		auto textFormat{ CreateTextFormat(fontName, static_cast<float>(fontSize), textAlignment, paragraphAlignment) };
+		auto textFormat{ CreateTextFormat(fontName, fontSize, textAlignment, paragraphAlignment) };
 		Renderer::ctx->DrawText(
 			text.data(),
 			static_cast<UINT32>(text.size()),
